Add PRODUCT_FILE & SERVER_PRODUCT_FILE params to incubator jenkins files
Add a stage to overwrite the tracing.product file if the PRODUCT_FILE
string parameter is defined, and respectively overwrite
traceserver.product if the SERVER_PRODUCT_FILE string parameter is
defined.

The string value should be the relative path to the product file from
the root of the product plug-ins.

Change-Id: Id85772ffcc0438ddb12db2af2c07a0ad5db45b8f
Signed-off-by: Bernd Hufmann <bernd.hufmann@ericsson.com>