[tracecompass-test.Jenkinsfile] First draft of notarization...
of mac .dmg packages (TraceCompass RPC package for mac)

This commit adds the necessary to notarize the mac .dmg packages,
as a test or example. This is exploratory and will require a follow
up, probably to refine the approach, and once it works, to add it
in the proper place (JenkinsFile / Jenkins job).

As noted in the code, this is heavily inspired by an Eclipse EPP
script and Jenkins job configuration. See code for details.

Signed-off-by: Marc Dumais <marc.dumais@ericsson.com>
Change-Id: Ib97133b16e14bbc79a3eb32116ad2b028d2d4183