[RCP][mac] add Jenkins stage for notarization
This stage is only executed if the Jenkins parameter "NOTARIZE_MAC_RCP",
that we introduce in this patch, is set to true. The default value is
false, and needs to be overridden in specific Jenkins jobs that really
require it (notarization is a limited resource).

Also, enhanced the notarization script to permit more than one package
being notarized at the same time and separately log the output of each
notarization for eventual troubleshooting.

Note: Notatization requires that mac .dmg packages be first generated
for the rcp product. The necessary is already in place, in the main
trace compass repo, and can be triggered in Jenkins jobs by adding the
maven argument "-Prcp-DMG-package-mac" in the appropriate field for
parameter "MAVEN_ARGS".

Signed-off-by: Marc Dumais <marc.dumais@ericsson.com>
Change-Id: I9356718c1e2d1a8551f36e751f806ebff50dc706