Remove setting of NOTARIZE_MAC_RCP in relevant jenkins files.
It seems that the setting of default value in the jenkins file take
precedence of the setting of the value in the jenkins job. Hence,
notarization has not been executed.

Implement parameter NOTARIZE_MAC_RCP the same way than other flags, for
example DEPLOY_RCP, and don't set it in the jenkinsfile. The jenkins job
configuration can take care of setting the value and provide a default
setting.

Change-Id: I1b552f1585d1e53f4b327454eb91332037a7f024
Signed-off-by: Bernd Hufmann <bernd.hufmann@ericsson.com>